moved       size_t totalMem; back to       unsigned int totalMem; because it wouldn't compile otherwise. Which one to use; should I have casted in order to compile or using unsigned int is fine?
git-svn-id: https://icl.cs.utk.edu/svn/magma/branches/sc_release@219 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ src/auxiliary.cpp
@@ -57,7 +57,7 @@ void printout_devices( )
   for( int idevice = 0; idevice < ndevices; idevice++ )
     {
       char name[200];
-      size_t totalMem;
+      unsigned int totalMem;
       int clock;
       CUdevice dev;
